+++ apps/client/src/routes/(menu)/multiplayer/[slug]/(waiting-room)/CurrentPlayers.svelte
@@ -1,24 +1,26 @@
 <script lang="ts">
-	import { roomStore } from '$lib/stores/MultiplayerStore';
+	import { roomStateStore, roomStore } from '$lib/stores/MultiplayerStore';
+	import type { PlayerState } from '@jmischler72/shared';
 
-	export let players: Map<string, boolean>;
+	$: players = $roomStateStore ? ($roomStateStore.players as Map<string, PlayerState>) : new Map<string, PlayerState>();
 </script>
 
 <div class="flex flex-col items-center justify-center gap-6 p-8">
 	<div class="flex flex-col items-center justify-center">
 		<ul class="list-none">
-			{#each players.keys() as player}
+			{#each players as [key, player]}
 				<li class="flex flex-row gap-2">
 					<h1
 						class="mb-2 rounded border-solid border-gray-400 bg-gray-600 p-2"
-						class:border-2={player === $roomStore?.sessionId}
+						class:border-2={key === $roomStore?.sessionId}
 					>
-						{player}
+						{key}
 					</h1>
-					{#if players.get(player)}
+					{#if player.ready}
 						<span class="mb-2 rounded border-solid border-gray-400 bg-gray-600 p-2 text-green-100">✔</span>
 					{/if}
-				</li>{/each}
+				</li>
+			{/each}
 		</ul>
 	</div>
 	{#if players.size < 2}