+++ apps/server/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.1.10 (2024-05-06)
+
+**Note:** Version bump only for package @jmischler72/server
+
+## 1.1.9 (2024-05-06)
+
+**Note:** Version bump only for package @jmischler72/server
+
+## 1.1.8 (2024-05-06)
+
+**Note:** Version bump only for package @jmischler72/server
+
 ## 1.1.7 (2024-04-30)
 
 **Note:** Version bump only for package @jmischler72/server
@@ -11,7 +23,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Bug Fixes
 
-* **client:** sync remove player with server ([ac9883e](https://github.com/jmischler72/tetrarena/commit/ac9883e591ee2e592f3b9ece4397411144a2168b))
+- **client:** sync remove player with server ([ac9883e](https://github.com/jmischler72/tetrarena/commit/ac9883e591ee2e592f3b9ece4397411144a2168b))
 
 ## 1.1.5 (2024-04-26)
 
@@ -21,7 +33,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Bug Fixes
 
-* **shared:** rename packages ([dc6527d](https://github.com/jmischler72/tetrarena/commit/dc6527d165dcb5d9d68f9eb3951bea06da1b37cb))
+- **shared:** rename packages ([dc6527d](https://github.com/jmischler72/tetrarena/commit/dc6527d165dcb5d9d68f9eb3951bea06da1b37cb))
 
 ## 1.1.3 (2024-04-26)